Moved data classes into new package named UI, those data classes are used in the UI, top layer of the app. And added new data class for the DetailsCompose.kt file
