+++ app/src/main/java/com/krakert/tracker/database/CryptoCacheDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.krakert.tracker.models.Coin
+import com.krakert.tracker.models.responses.Coin
 
 @Dao
 interface CryptoCacheDao {
