+++ app/src/main/java/com/krakert/tracker/database/CyptoDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.krakert.tracker.models.Coin
+import com.krakert.tracker.models.responses.Coin
 import com.krakert.tracker.models.database.Converters
 
 @Database(entities = [Coin::class], version = 1, exportSchema = false)
