--- app/src/main/java/com/krakert/tracker/models/Currency.kt
@@ -1,7 +1,0 @@
-package com.krakert.tracker.models
-
-enum class Currency(val nameFull: List<String>) {
-    USD(listOf("Dollar","$")),
-    EUR(listOf("Euro", "€")),
-    GBP(listOf("Pound", "£"))
-}
