+++ app/src/main/java/com/krakert/tracker/models/ui/Currency.kt
@@ -1,0 +1,7 @@
+package com.krakert.tracker.models.ui
+
+enum class Currency(val nameFull: List<String>) {
+    usd(listOf("Dollar","$")),
+    eur(listOf("Euro", "€")),
+    gbp(listOf("Pound", "£"))
+}
